add graficas

+++ Proyecto Integrador/Proyecto Integrador/Modelos/Cotizaciones/PuntoTerreno.cs
@@ -2,11 +2,11 @@
 
 public class PuntoTerreno
 {
-    public decimal X { get; set; }
-    public decimal Y { get; set; }
-    public decimal Z { get; set; }
+    public double X { get; set; }
+    public double Y { get; set; }
+    public double Z { get; set; }
 
-    public PuntoTerreno(decimal x, decimal y, decimal z)
+    public PuntoTerreno(double x, double y, double z)
     {
         X = x;
         Y = y;
